requirements.txt
asgiref==3.5.0
distlib==0.3.4
Django==4.0.3
djangorestframework==3.13.1
filelock==3.6.0
mysqlclient==2.1.0
Pillow==9.0.1
platformdirs==2.5.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.13.3
virtualenvwrapper-win==1.2.7
